pyproject.toml
[project]
name = "streamlit-perspective"
version = "0.0.2"
description = "Perspective components for Streamlit apps."
readme = "README.md"
authors = [{ name = "Manuel Šarić", email = "manujelko@gmail.com" }]
requires-python = ">=3.10"
dependencies = ["streamlit>=1.51"]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
wheel-exclude = [
  "node_modules",
  "frontend/src",
  "frontend/package.json",
  "frontend/package-lock.json",
  "frontend/tsconfig.json",
  "frontend/vite.config.ts",
]

[project.optional-dependencies]
examples = [
  "perspective-python>=4.1.1",
  "tornado>=6.0",
]

[dependency-groups]
dev = [
  "ruff>=0.15.0",
  "ty>=0.0.16",
]
